+++ src/data-load.ts
@@ -7,12 +7,14 @@ import {
   SearchFile,
   Users,
 } from "./interfaces.js";
+import { UserNames } from "./user-names.js";
 import {
   CHANNELS_DATA_PATH,
   EMOJIS_DATA_PATH,
   getChannelDataFilePath,
   SEARCH_DATA_PATH,
   USERS_DATA_PATH,
+  USER_NAMES_DATA_PATH,
 } from "./config.js";
 import { retry } from "./retry.js";
 import { readJsonArraySync, readSearchDataSync } from "./big-json.js";
@@ -66,6 +68,10 @@ export async function getUsers(): Promise<Users> {
 
 export async function getEmoji(): Promise<Emojis> {
   return getFile<Emojis>(EMOJIS_DATA_PATH, {});
+}
+
+export async function getUserNames(): Promise<UserNames> {
+  return getFile<UserNames>(USER_NAMES_DATA_PATH, {});
 }
 
 export async function getChannels(): Promise<Array<Channel>> {